move server id function to its own file in testdb module
