+++ server/testdb/postgres.go
@@ -8,7 +8,6 @@ import (
 	"github.com/golang-migrate/migrate/v4"
 	"github.com/golang-migrate/migrate/v4/database/postgres"
 	_ "github.com/golang-migrate/migrate/v4/source/file"
-	"github.com/kubeshop/tracetest/server/config"
 	"github.com/kubeshop/tracetest/server/id"
 )
 
@@ -61,44 +60,6 @@ func (p *postgresDB) ensureLatestMigration() error {
 	}
 
 	return nil
-}
-
-func (td *postgresDB) ServerID() (id string, isNew bool, err error) {
-	isNew = false
-	id = ""
-
-	err = td.db.
-		QueryRow(`SELECT id FROM "server" LIMIT 1`).
-		Scan(&id)
-	if err != nil && err != sql.ErrNoRows {
-		err = fmt.Errorf("could not get serverID from DB: %w", err)
-		return
-	}
-
-	if id != "" {
-		return
-	}
-
-	// no id, let's creat it
-	isNew = true
-	id = config.GetMachineID()
-	id = id[:10] // limit lenght to avoid issues with GA
-
-	stmt, err := td.db.Prepare(`INSERT INTO "server" (id) VALUES ($1)`)
-	if err != nil {
-		err = fmt.Errorf("could not prepare stmt: %w", err)
-		return
-	}
-	defer stmt.Close()
-
-	_, err = stmt.Exec(id)
-	if err != nil {
-		err = fmt.Errorf("could not save serverID into DB: %w", err)
-		return
-	}
-
-	return
-
 }
 
 func (td *postgresDB) Drop() error {
